[IMP] Reorganize wizards and date_tools
