+++ coop_shift/__manifest__.py
@@ -30,20 +30,13 @@
         'data/report_paperformat.xml',
         'data/email_template.xml',
         'views/action.xml',
-        'wizard/create_shifts_wizard_view.xml',
-        'wizard/update_shifts_wizard_view.xml',
-        'wizard/add_template_registration_view.xml',
-        'wizard/replace_registration_view.xml',
         'views/menu.xml',
-        'wizard/report_timesheet_wizard_view.xml',
-        'wizard/report_wallchart_wizard_view.xml',
         'views/product_view.xml',
         'views/shift_mail_view.xml',
         'views/shift_registration_view.xml',
         'views/shift_template_registration_view.xml',
         'views/shift_template_registration_line_view.xml',
         'views/shift_shift_view.xml',
-        'wizard/create_shifts_wizard_view2.xml',
         'views/shift_template_view.xml',
         'views/shift_type_view.xml',
         'views/shift_ticket_view.xml',
@@ -53,13 +46,20 @@
         'views/view_shift_extension_type.xml',
         'views/view_shift_leave.xml',
         'views/view_shift_leave_type.xml',
-        'views/view_shift_leave_wizard.xml',
         'views/report_layout_custom.xml',
         'views/shift_report.xml',
         'views/shift_report2.xml',
         'views/report_timesheet.xml',
         'views/report_timesheet_inherited.xml',
         'views/report_wallchart.xml',
+        'wizard/create_shifts_wizard_view.xml',
+        'wizard/update_shifts_wizard_view.xml',
+        'wizard/add_template_registration_view.xml',
+        'wizard/replace_registration_view.xml',
+        'wizard/view_shift_leave_wizard.xml',
+        'wizard/create_shifts_wizard_view2.xml',
+        'wizard/report_timesheet_wizard_view.xml',
+        'wizard/report_wallchart_wizard_view.xml',
     ],
     'demo': [
         'demo/res_groups.xml',
