+++ coop_shift/date_tools.py
