+++ coop_shift/model/__init__.py
@@ -15,6 +15,4 @@ from . import res_partner
 from . import product
 from . import shift_leave
 from . import shift_leave_type
-from . import shift_leave_wizard
 from . import shift_counter_event
-from . import date_tools
