+++ coop_shift/wizard/add_template_registration.py
@@ -5,7 +5,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from odoo import models, fields, api
-from ..model.date_tools import conflict_period
+from ..date_tools import conflict_period
 
 STATES = [
     ('cancel', 'Cancelled'),
