+++ coop_shift/wizard/view_shift_leave_wizard.xml
